cast to bouble to avoid ambiguaty

+++ win32ss/gdi/ntgdi/arc.c
@@ -122,8 +122,8 @@ IntArc( DC *dc,
 
     CenterX = (RectBounds.right + RectBounds.left) / 2;
     CenterY = (RectBounds.bottom + RectBounds.top) / 2;
-    AngleEnd   = atan2((RectSEpts.bottom - CenterY), RectSEpts.right - CenterX)*(360.0/(M_PI*2));
-    AngleStart = atan2((RectSEpts.top - CenterY), RectSEpts.left - CenterX)*(360.0/(M_PI*2));
+    AngleEnd   = atan2((double)(RectSEpts.bottom - CenterY), (double)(RectSEpts.right - CenterX))*(360.0/(M_PI*2));
+    AngleStart = atan2((double)(RectSEpts.top - CenterY), (double)(RectSEpts.left - CenterX))*(360.0/(M_PI*2));
 
     /* Edge Case: Check if the start segments overlaps(is equal) the end segment */
     if (AngleEnd == AngleStart)
